updated settings reducer
