+++ app/components/sidemenu/constants.js
@@ -1,1 +1,3 @@
-export const HANDLE_MENU_CHANGE = 'HANDLE_MENU_CHANGE'
+export const actions = {
+	HANDLE_MENU_CHANGE: 'HANDLE_MENU_CHANGE'
+}
