+++ app/package.json
@@ -1,18 +1,18 @@
 {
-  "name": "marks",
-  "productName": "marks",
-  "version": "1.3.0",
-  "description": "This app gives teachers the ability to keep track of marks.",
-  "main": "./main.prod.js",
-  "author": {
-    "name": "Henry Cuffy",
-    "email": "hcuffy7@gmail.com",
-    "url": "https://github.com/hcuffy/Marks"
-  },
-  "scripts": {
-    "electron-rebuild": "node -r ../internals/scripts/BabelRegister.js ../internals/scripts/ElectronRebuild.js",
-    "postinstall": "yarn electron-rebuild"
-  },
-  "license": "MIT",
-  "dependencies": {}
+	"name": "marks",
+	"productName": "marks",
+	"version": "1.3.2",
+	"description": "This app gives teachers the ability to keep track of marks.",
+	"main": "./main.prod.js",
+	"author": {
+		"name": "Henry Cuffy",
+		"email": "hcuffy7@gmail.com",
+		"url": "https://github.com/hcuffy/Marks"
+	},
+	"scripts": {
+		"electron-rebuild": "node -r ../internals/scripts/BabelRegister.js ../internals/scripts/ElectronRebuild.js",
+		"postinstall": "yarn electron-rebuild"
+	},
+	"license": "MIT",
+	"dependencies": {}
 }
