+++ package.json
@@ -1,7 +1,7 @@
 {
 	"name": "marks",
 	"productName": "Marks",
-	"version": "1.3.0",
+	"version": "1.3.2",
 	"description": "This app gives teachers the ability to keep track of marks.",
 	"scripts": {
 		"build": "concurrently \"yarn build-main\" \"yarn build-renderer\"",
